Issue #65: Minor changes: partially restructure dedicated view and helper modules in views folder. Optimized imports.
